By default, the pull is now done directly instead of creating a branch
A branch was created, and pull was done in it, and a merge had to be done after by hand.
Since I do the review in Github, I prefer to do the merge directly when I now that it can
be done without conflicts. And this is the case most of time for me.

You can keep the old behavior by using the -b option, which will create a branch.